Create modules and separate files
